# specs/001-ai/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: 智能標書產生器系統 API
  description: 提供公司資料管理、範本管理、AI內容生成和標書編輯匯出功能的RESTful API
  version: 1.0.0
  contact:
    name: 標書產生器團隊
    email: support@proposal-generator.com
servers:
  - url: http://localhost:3000/api/v1
    description: 開發環境
  - url: https://api.proposal-generator.com/v1
    description: 生產環境
paths:
...
  /team-members:
    get:
      tags: [Team]
      summary: 取得團隊成員列表
      security:
        - bearerAuth: []
      parameters:
        - name: is_key_member
          in: query
          schema:
            type: boolean
          description: 篩選關鍵人員
      responses:
        '200':
          description: 成功取得團隊成員列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamMember'

    post:
      tags: [Team]
      summary: 新增團隊成員
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberCreate'
      responses:
        '201':
          description: 成功新增團隊成員
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
  /team-members/{id}:
    put:
      tags: [Team]
      summary: 更新團隊成員
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberUpdate'
      responses:
        '200':
          description: 更新成功
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Team]
      summary: 刪除團隊成員
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 刪除成功
        '404':
          $ref: '#/components/responses/NotFound'

  # 實績案例 API
...
  /proposals:
    get:
      tags: [Proposals]
      summary: 取得標書列表
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProposalStatus'
      responses:
        '200':
          description: 成功取得標書列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Proposal'

    post:
      tags: [Proposals]
      summary: 建立新標書
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalCreate'
      responses:
        '201':
          description: 成功建立標書
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposal'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TeamMember:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        title:
          type: string
        department:
          type: string
        education:
          type: string
        experience:
          type: string
        expertise:
          type: string
        photo_url:
          type: string
          format: uri
        is_key_member:
          type: boolean
        display_order:
          type: integer
    TeamMemberCreate:
      type: object
      required: [name, title]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        title:
          type: string
        department:
          type: string
        education:
          type: string
        experience:
          type: string
        expertise:
          type: string
        is_key_member:
          type: boolean
          default: false
    TeamMemberUpdate:
      allOf:
        - $ref: '#/components/schemas/TeamMemberCreate'
        - type: object
          properties:
            display_order:
              type: integer
              minimum: 0
...
    Proposal:
      type: object
      properties:
        id:
          type: string
        proposal_name:
          type: string
        template_id:
          type: string
        status:
          $ref: '#/components/schemas/ProposalStatus'
        word_count:
          type: integer
        generated_with_ai:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProposalStatus:
      type: string
      enum: [draft, in_review, completed, submitted]
    ProposalCreate:
      type: object
      required: [proposal_name, template_id]
      properties:
        proposal_name:
          type: string
          minLength: 2
          maxLength: 255
        template_id:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
  responses:
...
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...